Extend README with information about how to contribute.
Added a contribution section that touches on using a normal github
workflow with pull requests, issues and so on.
Also mention that all types of issues are welcome, security problems
are not supposed to be reported in an alternativ path.

These changes are part of qualification for the CII Best practices
badge.