Add automatic dependency checking using dependabot.
Adding a configuration file for checking the composer
dependencies every Thursday. It will open a PR if needed be,
maximum 2 PRs can be opened - every two weeks an action must be
made. Should be possible to cope with.

These changes are part of qualification for the CII Best practices
silver badge.