Update data loader behaviour for integer ID-based "associative" keys
The prior logic for determining whether a data loader result was index-based or
contained associative keys was flawed. This could lead to a problematic
situation in which integer IDs were used as "associative" keys while posing the
risk of "matching" an index:

Index-based array: `[0 => ..., 1 => ..., 2 => ...]`

Integer ID-based array: `[2 => ..., 1 => ..., 3 => ...]`

Since the previous logic performed this check for each iteration instead of once
for the entire result, it interpreted the second item in the integer ID-based
array above as index-based (since ID 1 corresponds to index 1).

This commit updates the logic to use `array_is_list()` to prevent this flaw.